# api-specs/openapi.yaml
openapi: 3.0.0
info:
  title: WhatsApp API Server
  version: 1.0.0
  description: API server for interacting with WhatsApp Web services
servers:
  - url: /api/v1
    description: API v1 Base URL
paths:
  /health:
    get:
      summary: Health check endpoint
      servers:
        - url: /api
          description: API Health URL
      responses:
        '200':
          description: Returns ok.
  /auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phoneNumber
                - username
                - password
              properties:
                phoneNumber:
                  type: string
                  description: User's phone number
                username:
                  type: string
                  description: Username for registration
                password:
                  type: string
                  description: Password for registration
      responses:
        '200':
          description: User registered successfully
  /auth/connect:
    post:
      summary: Connect to WhatsApp
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Connection status and QR code if available
  /auth/connect/status:
    get:
      summary: Get connection status
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Whether the connection is ready
  /auth/connect/qr:
    get:
      summary: Get QR code for connection
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: QR code string
  /auth/connect/pairing-code:
    get:
      summary: Get pairing code for connection
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Pairing code string
  /stream:
    get:
      summary: Connect to message stream (SSE)
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Event source stream
  /send/typing:
    post:
      summary: Send typing indicator
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phoneNumber
              properties:
                phoneNumber:
                  type: string
      responses:
        '200':
          description: Success
        '400':
          description: Phone number required
  /send/text:
    post:
      summary: Send text message
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phoneNumber
                - message
              properties:
                phoneNumber:
                  type: string
                message:
                  type: string
      responses:
        '200':
          description: Success
        '400':
          description: Phone number and message required
  /send/file:
    post:
      summary: Send media file
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - phoneNumber
                - file
              properties:
                phoneNumber:
                  type: string
                caption:
                  type: string
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Success
        '400':
          description: Phone number and file required
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key