# src/routes/v1/whatsappRoutes.ts
import express from "express";
import { whatsappService } from "../../lib/whatsapp.js";
import multer from "multer";
import authMiddleware from "../../middleware/auth.js";
import { logger } from "../../lib/logger.js";
import validate from "../../middleware/validate.js";

const upload = multer({ storage: multer.memoryStorage() });
const router = express.Router();

router.get('/stream', authMiddleware, (req, res) => {
    res.writeHead(200, {
        'content-type': 'text/event-stream',
        'cache-control': 'no-cache',
        'connection': 'keep-alive'
    });

    res.write(`data: {"status": "connected"}\n\n`);

    const listener = (apiKey: string, message: any) => {
        if (apiKey !== res.locals.apiKey) {
            return;
        }
        res.write(`event:message\ndata: ${JSON.stringify(message)}\n\n`);
    }

    whatsappService.on('message', listener);

    req.on('close', () => {
        logger.info('Client disconnected');
        whatsappService.off('message', listener);
        res.end();
    })
});

router.post('/send/typing', authMiddleware, validate({ body: { phoneNumber: "required|phoneNumber" } }), async (req, res) => {
    const apiKey = res.locals.apiKey;
    const { phoneNumber } = req.body;
    const success = await whatsappService.sendTyping(apiKey, phoneNumber);

    if (!success) {
        return res.status(400).json({
            success: false,
            error: {
                code: 400,
                message: "Failed to send typing indicator"
            }
        });
    }

    return res.json({
        success: true,
        data: { message: "Typing sent successfully" }
    });
});

router.post('/send/text', authMiddleware, validate({ body: { phoneNumber: "required|phoneNumber", message: "required|message" } }), async (req, res) => {
    const apiKey = res.locals.apiKey;
    const { phoneNumber, message } = req.body;
    const success = await whatsappService.sendMessage(apiKey, phoneNumber, message);
    if (!success) {
        return res.status(400).json({
            success: false,
            error: {
                code: 400,
                message: "Failed to send message"
            }
        });
    }
    return res.json({
        success: true,
        data: { message: "Message sent successfully" }
    });
});

router.post('/send/file', authMiddleware, upload.single('file'), validate({ 
    body: { 
        phoneNumber: "required|phoneNumber",
        caption: "string"
    }, 
    file: "required|file"
}), async (req, res) => {
    const apiKey = res.locals.apiKey;
    const { phoneNumber, caption } = req.body;
    const file = req.file!;
    logger.info({ body: req.body, file: req.file?.originalname });
    const success = await whatsappService.sendMediaFile(apiKey, phoneNumber, file, caption);
    if (!success) {
        return res.status(400).json({
            success: false,
            error: {
                code: 400,
                message: "Failed to send file"
            }
        });
    }
    return res.json({
        success: true,
        data: { message: "File sent successfully" }
    });
});

export default router;